Made constants and handler packages internal in the lutron binding
