+++ addons/binding/org.openhab.binding.lutron/META-INF/MANIFEST.MF
@@ -7,8 +7,6 @@ Bundle-RequiredExecutionEnvironment: JavaSE-1.8
 Bundle-SymbolicName: org.openhab.binding.lutron;singleton:=true
 Bundle-Vendor: openHAB
 Bundle-Version: 2.4.0.qualifier
-Export-Package: org.openhab.binding.lutron,
- org.openhab.binding.lutron.handler
 Import-Package: com.google.common.collect,
  com.thoughtworks.xstream,
  gnu.io,
@@ -26,8 +24,6 @@ Import-Package: com.google.common.collect,
  org.eclipse.smarthome.core.thing.binding.builder,
  org.eclipse.smarthome.core.thing.type,
  org.eclipse.smarthome.core.types,
- org.openhab.binding.lutron,
- org.openhab.binding.lutron.handler,
  org.osgi.framework,
  org.slf4j
 Service-Component: OSGI-INF/*.xml
