+++ addons/binding/org.openhab.binding.lutron/src/main/java/org/openhab/binding/lutron/internal/radiora/handler/SwitchHandler.java
@@ -12,7 +12,7 @@ import org.eclipse.smarthome.core.library.types.OnOffType;
 import org.eclipse.smarthome.core.thing.ChannelUID;
 import org.eclipse.smarthome.core.thing.Thing;
 import org.eclipse.smarthome.core.types.Command;
-import org.openhab.binding.lutron.LutronBindingConstants;
+import org.openhab.binding.lutron.internal.LutronBindingConstants;
 import org.openhab.binding.lutron.internal.radiora.config.SwitchConfig;
 import org.openhab.binding.lutron.internal.radiora.protocol.LocalZoneChangeFeedback;
 import org.openhab.binding.lutron.internal.radiora.protocol.RadioRAFeedback;
